fix: move select/stop to coincide w/ running job
The stop_queue() test revealed a logic/flow problem when
the process_job task receives a Stop message.

The tokio::select!{} was in the wrong place. It was selecting between
the stop message and the job_added message.  Instead it needs to select
between the running job and the stop message.

This commit makes that change and now the stop_queue() test passes.