style: Reverse removal records prior to applying
Stylistic change as opposed to logical refactoring because, it turns,
it doesn't matter: the indices inside of each `Chunk` are sorted; and
the resulting authentication paths are independent of the order of the
modifications. But it's still a good idea to follow the same procedure
as elsewhere.

Also: integrate reviewer comments.