fix: avoid panic in TransactionDetails::new()
closes #586.

modifies `TransactionDetails::new()` to return `Self` instead of
an anyhow::Result<Self> and removes the lines that caused the panic.

callers can call TransactionDetails::validate() if they wish for much more
comprehensive validation.

This has the nice benefit that we then were able to remove the `Failed` variant
of `CreateTxError` that was for handling anyhow::Error.

We go further to also return a new `RecordTransactionError` from
`GlobalStateLock::record_transaction()` which enables removal of
the catch-all `SendError::Failed`.