fix: validate TransactionDetails
This change ensures that TransactionDetails are validated
when GlobalStateLock::record_transaction() is called and can catch validation
errors for mock-proof using networks that otherwise might slip through.

It is not strictly necessary to validate TransactionDetails when the network
uses real proofs because a valid proof indicates the Tx is valid and also the
TransactionDetails, so long as it matches the Tx.

However:
 1. A network that uses mock proofs should validate the TransactionDetails
because a "valid" mock-proof can be obtained with invalid TransactionDetails.

 2. The enum error returned from TransactionDetails provides more insight into
what is wrong than the boolean value from Transaction::verify_proof()

 3. It is preferable to use the same logic whether the network uses mock proofs
or not.

For these reasons, TxCreationArtifacts::verify() now validates the
TransactionDetails.