bench(wallet): Add benchmark of wallet state update
Add a benchmark that measures the time it takes to update a wallet
monitoring 100 UTXOs with a block in which an additional 100 UTXOs are
received.

This benchmark is added because power users report slow state-updating
times, and the *very* slow (up to 5 seconds) state updates have been
pinpointed to being in the state-updater by block for the wallet.