perf: Only flush once when applying multiple blocks
Move the flushing of databases from `GlobalState`'s two interfaces for
storing blocks, `set_new_tip` and `store_block_not_tip`, to the caller.
This has the effect that when the state is updated with multiple blocks,
the flushing only needs to happen once.

Outside of tests, this only happens three places:
- one instance in main loop's `handle_self_guessed_block` where an
   explicit call to `flush_databases` is added in this commit. In this
   case, there is only ever one block that's applied, so this commit
  gives no speedup there.
- two instances in main_loop's `PeerTaskToMain::NewBlocks`, one where
  the new blocks are canonical (more PoW than current tip), and one
  where they are not canonical (not more PoW than current tip). In both
  of these cases, an explicit call to flush databases have been added
  outside of a loop over blocks, which will speedup the process when
  multiple blocks are received from a peer.