fix: Account for reorganization in bootstrap source
When bootstrapping blocks, allow for reorganizations to have taken place
on the node that produced the `blocks` directory used in the
bootstrapping process.

The predecessor of the block read from the `blk` files in the `blocks`
directory may not be the previous block read. This commit fetches the
correct predecessor if it was not the previously handled block. The
correct predecessor is fetched from the node's archival state, as it is
assumed to be available.