refactor(GlobalState): improve encapsulation of constructor
Change function signature and logic of `GlobalState::new` to move more
of the database initialization into this constructor. Previously, this
database initialization happened directly in lib.rs's `initialize`.

Also allow caller of this constructor to specify the genesis block, as
this will widen our testing capabilities somewhat.

Big diff because this changes many tests. But the diff of non-testing,
mainline, code is quite small.