feat: Version consensus programs
The upcoming hard fork requires two sets of consensus rules and thus
two sets of consensus programs. This commit provides the scaffolding
for these changes. It is a squash of plenty commits. The important
features are summarized below.

 - Pass parameter `network` to plenty of functions where it is needed.
 - Replicate tests across multiple consensus rule sets, typically
   using macros.
 - Add enum `ConsensusRuleSet` which captures *differences* in the
   consensus rules across forks. Expanded below.

feat!(`ConsensusRuleSet`): Introduce and use struct `ConsensusRuleSet`

In order to enforce different sets of consensus rules based on the
block height at which the change is triggered, it became necessary to
pass the block height to the various validation functions. However,
this parameter is at odds with the convention that validity is
independent of block height. (That would be *confirmability*, which is
treated separately.) This confusion is now avoided by passing a
`ConsensusRuleSet` instead. The fact that you need the block height to
populate a variable of type `ConsensusRuleSet` is an artifact of
having forks triggered intentionally at certain block heights and does
not need to pollute the signature of validation methods.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>
Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>