style: Consistently use the same way to `div_rem`
As part of a PR revision, I suggested to either use `div_rem_euclid` or
a custom `div_rem` closure, providing examples of how to use either in
order to let the author choose between them. This was admittedly subtly
intervowen in the rest of the function. For consistency reasons, only
one of the two suggested methodes should be used – and now, it is.

changelog: ignore