bench: Add a benchmark for wallet-state updating
When syncing on a node with many UTXOs, the wallet state update is **by
far** the slowest part of the state update. So having a benchmark for
this critical step is critical.

Adds a module `bench_helpers` to encapsulate code needed for benchmarks.
Since benchmarks are not part of the main crate, they can only use items
that are `pub`. So this `bench_helpers` module prevents us from marking
half the repo as `pub`, as all functions needed for benchmarks can be
neatly encapsulated here.