feat(cli_args): Add option `tx_proof_upgrading`
Start implementing parts of three-step mining as outlined in:
https://talk.neptune.cash/t/thoughts-on-the-mempool-issue-and-upgrading/52/26.

This is done by adding the `tx_proof_upgrading` option to the CLI
arguments which can be used to specify whether to participate in
transaction proof upgrading of 3rd party transactions.

With this change the scheduled task of transaction-proof upgrading only
happens when this value is explicitly set. Previously, powerful enough
machines would automatically participate in proof upgrading.