fix: make regtest launch-date same as others
The regtest network previously used a launch date that was normalized to the
previous week. This does not appear to serve any purpose presently as all tests
pass without it.

The regularly changing launch_date() makes it harder to write deterministic test
cases for regtest, which regtest is supposed to facilitate.

It further causes a hidden test failure for the recently updated test
genesis_block_hasnt_changed().  That test would work for 1 week after the
"known" genesis block hash is determined, then begin failing afterward.

The simplest way to make this test work and avoid special cases is to fix the
regtest launch_date in time, as the other networks are.

This commit also changes Network::launch_date() from pub(crate) to pub.  All
other methods in `impl Network` are pub and there doesn't seem to be any
reason this one should not be.