feat(archival_state): Add function to get historical MS data
Add a function to get an old mutator set accumulator from its hash. The
mutator set returned will alwaays be one that aligns with a block, so
this function only allows for finding mutator sets that match with
historical blocks.

This functionality can be used to define "update" jobs where single
proofs are updated to be valid under a new mutator set.