feat(mempool): Get most valuable update job
Add a new endpoint to mempool returning the most valuable transaction
for proof updating. Uses the transaction's associated `UpgradePriority`
to return the favored transaction for (single proof) updating. This new
endpoint is in the same family as `preferred_proof_collection` and
`preferred_single_proof_pair` which are also used by the proof upgrader
to find potential proof work.

Also use this new endpoint in the proof upgrading logic.