chore: Gracefully catch errors resulting from invalid blocks
The peer loop should act as a robust firewall against invalid blocks.
Nevertheless, it is worth ensuring that handlers behind that firewall
cannot crash even for invalid blocks. This commit replaces
`.unwrap()`s and `.expect(..)`s with question marks, and wraps the
return type of some handlers in a `Result` where necessary.

Thanks to @skaunov for identifying the issues.