style: Suppress funky linter warnings
Suppresses linter warning `clippy::explicit_deref_methods` for test
modules containing `test_strategy` proptests with a `strategy`
directive for ranges. Specifically, clippy (erroneously) complains
about lines like this:

```
    |
381 |             #[strategy(3usize..20)] n: usize,
    |                        ^^^^^^^^^^ help: try: `&**3usize..20`
    |
```

Clippy is complaining about some code emitted by the `test_strategy`
procedural macro, that the end-developer does not see. This funky is
probably resolved in a future version of clippy or of `test_strategy`.
When it is resolved, the lines of the form

```rust
```

can all be removed.