refactor(mempool): Only return pw-backed txs for MS-updating
When the mempool processes a new block, use the presence of a primitive
witness to determine whether the transaction should be updated to be
valid under the new mutator set or not.

Previously, a composer would update all single-proof backed transactions
before starting mining again. Now, this high-priority transaction proof
upgrading (updating) only happens on transactions that were initiated
locally, i.e. transactions for which a primitive witness known. This
means that the mempool on its state update with a new block returns a
list of jobs describing how to update the primitive witness, proof
collection, or single to be valid under a new block. Previously, only
the single proof backed transactions where represented in the return type.

Since it's not possible to update a proof collection to be valid under a
new mutator set, we instead keep the primitive witness around (for
locally initiated transactions) and update the primitive witness first,
and then generate an updated proof collection from it. For the updating
of single proofs, we apply an "update" invocation of the SingleProof
program.