fix(mempool): Always allow mutator set data to be updated
When attempting to insert a transaction into the mempool we check if the
new transaction conflicts with existing transaction. If it does, we
check if it has a higher proof quality than its conflicts. Only if the
proof quality is higher does the new transaction replace the formed.
But! if the proof qualities are the same, we must check if the newly
inserted transaction represents a mutator set update. If it does, then
the new transaction must replace the former. This check was previously
only done for single proof backed transactions. With this commit, that
logic is also applied to ProofCollection and PrimitiveWitness backed
transactions.

The newly added test to `main_loop`, `tx_ms_updating` revealed this
deficiency. Thank you test!