refactor: Don't reset difficulty on `Testnet` network
This restores `Testnet` behavior to that prior to
1293adaff5df7be02382b867424fca2604692eac. Except that the initial test
net difficulty is now 10^6 and not 10^9. So this commit is both a
hardfork in relation to behavior prior to this commit, and a hardfork
prior to 1293adaff5df7be02382b867424fca2604692eac. In other words: This
commit introduces testnet v2 that we can hopefully get up and running
again.

The few people who were running test net prior to
1293adaff5df7be02382b867424fca2604692eac will have to clear their
testnet data directory which is `~/.local/share/neptune/testnet`
on Linux machines.