fix(peer_loop): Accept transactions with updated mutator sets
If a transaction with a specific mutator set is known, peer loop must
accept a new transaction if this new transaction represents a update of
the transaction already in the mempool. This fix is implemented by
changing the implementation of
`Mempool::contains_with_higher_proof_quality` which now checks the
mutator set digest of the transactions in question, and assumes that a
new mutator set value represents an updated transaction.