fix(mempool): Catch more events
Prior to this commit the mempool could in some instances remove transactions
without returning the associated events, leading to incorrect behavior
downstream in the wallet, as it uses the events for input selection.

This commit ensures that every time a transaction is added or removed
that an event is returned to the caller, and that the caller
(GlobalState) informs wallet of these events.