fix(mempool): Respect max size
Previously, we only measured the size of the pointer, not the size of
the mempool when checking if we needed to drop transactions. You
apparently need to dereference the pointer to get the size of the
object, as opposed to the size of the pointer.

Tests that would have caught this are added in the next commit.