refactor(mempool): Remove cap for number of transactions
Remove the functionality of the mempool to cap the number of
transactions in the mempool.

With the merger of 3f7fb11bdaf9c6b79819f86fef3d411b503ce11b and the
changes introduced in c85811e210256acbf2e63707e7b63f75d0f09600 and
in particular f70bccc22faf2e70b2a5e58c0c1a82ea59db30fd which allows
the composer control over the number of txs to merge, we no longer
have a good reason to cap the number of transactions in the
mempool, as there are no longer situations in which the presence of
mempool transactions causes a big slowdown for composers, or anyone else
for that matter.