feat(mining): Add CLI argument to set num merged transactions
Add a CLI argument to control the number of (single proof backed) transactions
that are merged in from the mempool. Previously this number was capped
at seven but with this commit, the number defaults to 1 since that's
what's in the interests of a composer. It's then up to the proof
upgraders to ensure that as many transactions as possible are merged
into one.