feat(mempool): Add new mempool field for preserving merge-inputs
When two transactions (a, b) are merged into transaction c, a and b are
dropped from the mempool. If the next block only contains transaction a,
then transaction b, which could be included in a block later, is lost.
This commit fixes this problem by introducing a cache for the mempool
such that b is restored to the mempool after a block with input a is set
as the new tip.

This closes #649.