refactor(`Transaction`): Add newtype for block transaction
This new type represents a transaction that can be included in a block.
Any transaction that has the merge bit set can be converted to a
`BlockTransaction` but the type is only instantiated in a composition
context.

Introduce and use test helper function `BlockTransaction::upgrade` which sends
`Transaction` -> `BlockTransaction` by setting the merge bit.

Use dedicated merging pipeline for block transactions

Co-authored-by: @sword_smith <thor@neptune.cash>
Co-authored-by: Alan Szepieniec <alan@neptune.cash>