refactor(`TransactionProofBuilder`): Make `ConsensusRuleSet` optional
`ConsensusRuleSet` is not required when constructing primitive witnesses or
proof collections. Therefore, it should not be a requirement but an
optional field.

Also: refactor the `from_witness` internal helper function by specializing it
into two functions, one for proof collection and one for single proofs. (The
pathway to produce a `PrimitiveWitness` from a `Cow<PrimitiveWitness>`
is eliminated because moot.)

Co-authored-by: sword-smith <thor@neptune.cash>
Co-authored-by: Alan Szepieniec <alan@neptune.cash>