refactor!(native_currency): Simplify check that half coinbase is timelocked
New rule: When coinbase is non-zero, verify that timelocked amount is at
least half of the output amount. This combined with a rule that
negative-fee transactions cannot be merged with coinbase transactions
should ensure correct amount is timelocked, according to the inflation
schedule.

A few tests are failing in native_currency but that's because of changed
assumptions, so the tests should be changed, not the main line code (at
least not because of those tests).