harden(collect_type_scripts): Bound num inputs/outputs and num coins
Set an upper limit for both number of input/outputs and for number of
coins per UTXO. Limit is set so high that it should never be reached in
any meaningful, constructive context. It is set in the expectation that
Triton VM might have continuations in the future, and it might be
possible to prove **very** big programs, in which case we want to ensure
that you can't somehow loop around in memory.