fix(`PrimitiveWitness`): Ensure presence of `NativeCurrency` hash
If absent, undetectable inflation is possible.

This fix changes nothing about the enforcement; it changes how the
witness is generated. It factors out a useful abstraction for computing
the list of type script hashes.

Also: fix(`PrimitiveWitness`): Avoid adding `TimeLock` type script
witnesses when coinbase is `Some(0)` in test case generator.

Note that uniformly random blocks, the block subsidy is zero with
overwhelming probability, leading to `Some(0)` coinbase. In this case,
however, no UTXOs need to be locked -- and so the `TimeLock` hash may
not be present either.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>