fix(`RemovalRecordList`): Fix packing of `RemovalRecord`s
This is a combination of 10 commits. Important points summarized below.

1. Factor out tricky MMR leaf count conversion

Going from the leaf count in the AOCL MMR to the leaf count in the
SWBFI MMR is tricky. This commit factors out the correct map into
a function called `aocl_to_swbfi_leaf_counts`. The function is
documented, to avoid repeating the mistake. This function is now
called in the places where the number of was previously computed
manually (sometimes correctly, sometimes incorrectly).

2. Fix off-by-one error in AOCL leaf count estimator and everywhere else

The correct relation between the number of leafs in the AOCL and the
number of leafs in the SWBFI is a constant factor *and* a constant
offset by one. This error was made in (and now fixed) in plenty of
places.

The new function `aocl_to_swbfi_leaf_counts`, added in this commit,
factors out this logic in one direction, but the estimator
fixed by this commit computes the map in the other direction. The
reason why it worked previously even though it was erroneous is because
it was consistent with the rest of the code, which fell prey to the same
pitfall.

3. Identify and reject inconsistent `RemovalRecordList`s

Add function `is_consistent` which checks if the `RemovalRecordList`
is consistent. It is possible to forge a malicious packed `Vec` of
`RemovalRecord`s such that unpacking will result in a (not consistent)
`RemovalRecordList` which is then cast into a unpacked `Vec` of
`RemovalRecord`s. The second step is rightly assumed to be infallible.
The `try_unpack` method now tests the consistency of the ephemeral
`RemovalRecordList` object and returns an error if it is inconsistent.

Consistency is for `RemovalRecordList` is defined relative to a set of
"observed chunk indices", which itself is obtained by filtering the
set of all absolute indices for lying in the inactive part of the SWBF
and mapping them to their corresponding chunk index. Specifically,
consistency for `RemovalRecordList` is defined as follows.
 1. The number of observed chunk indices agrees with the number of
    chunks in the data structure.
 2. The number of active trees (trees with chunks that are contained
    in this data structure) is less than or equal to the number of
    peaks in the SWBFI MMR. The latter number is determined based on
    the SWBFI leaf count, which in turn is inferred from the AOCL leaf
    count estimate. Furthermore, the number of authentication
    structures agrees with the number of peaks.
 3. For every tree, the observed length of the authentication
    structure matches the expected length of the authentication
    structure, where the expectation comes from deriving the
    authentication structure node indices for the leaf indices
    corresponding to the chunk indices living in that tree.

Consistency captures the total number of authentication structures, not
just the number of non-empty ones.

4. Use u64::MAX as dummy-delimiter for tree height

5. Account for more trees than chunks in encoding

Specifically: usean offset (`ENCODING_TREE_HEIGHT_OFFSET`) to indicate
that the `Chunk` should be ignored.

Also:
 - fix off-by-one error
 - improve doc strings
 - harden tests and add more

Co-authored-by: Alan Szepieniec <alan@neptune.cash>