refactor(`mutator_set`): Factor out tricky MMR leaf count conversion
Going from the leaf count in the AOCL MMR to the leaf count in the
SWBFI MMR is tricky. This commit factors out the correct map into
a function called `aocl_to_swbfi_leaf_counts`. The function is
documented, to avoid repeating the mistake. This function is now
called in the places where the number of was previously computed
manually (sometimes correctly, sometimes incorrectly).