fix(`RemovalRecordList`): Identify and reject inconsistent `RemovalRecordList`s
Add function `is_consistent` which checks if the `RemovalRecordList`
is consistent. It is possible to forge a malicious packed `Vec` of
`RemovalRecord`s such that unpacking will result in a (not consistent)
`RemovalRecordList` which is then cast into a unpacked `Vec` of
`RemovalRecord`s. The second step is rightly assumed to be infallible.
The `try_unpack` method now tests the consistency of the ephemeral
`RemovalRecordList` object and returns an error if it is inconsistent.

Consistency is for `RemovalRecordList` is defined relative to a set of
"observed chunk indices", which itself is obtained by filtering the
set of all absolute indices for lying in the inactive part of the SWBF
and mapping them to their corresponding chunk index. Specifically,
consistency for `RemovalRecordList` is defined as follows.
 1. The number of observed chunk indices agrees with the number of
    chunks in the data structure.
 2. The number of active trees (trees with chunks that are contained
    in this data structure) is less than or equal to the number of
    peaks in the SWBFI MMR. The latter number is determined based on
    the SWBFI leaf count, which in turn is inferred from the AOCL leaf
    count estimate. Furthermore, the number of authentication
    structures agrees with the number of peaks.
 3. For every tree, the observed length of the authentication
    structure matches the expected length of the authentication
    structure, where the expectation comes from deriving the
    authentication structure node indices for the leaf indices
    corresponding to the chunk indices living in that tree.