test: Fix test about empty tx with non-zero fee
This was one of the tests written to showcase the problem leading to the
reboot of the blockchain at block height 21310. This commit adapts the
test to be one that passes now after the bug has been fixed. The bug was
that the native currency type script was not run when the transaction
had zero inputs and outputs, meaning that bad fee amounts (meaning fee
amounts that allow for unplanned money printing) could be set and the
proof would still go through.