fix(`WalletState`): Do attempt to resync MUTXOs marked spent
After all, the resync may be triggered by a reorganization that
reverts the expenditure. In this case we want to have a valid
membership proof for the now-spendable-again UTXO.

Attempting to resync MUTXOs that are marked as spent should be
expected to fail most of the time, though, for the simple reason
that the spending transaction was *not* reverted by the reorg,
or because there was no reorg to begin with. For this reason, the
warning log message is suppressed when resyncing fails and the
UTXO was marked as spent.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>