feat(`GenerationReceivingAddress`): Implement `BFieldCodec`
The difficulty revolves around the field `encryption_key` which is
of type `lattice::kem::PublicKey`, lives in repository twenty-first,
and does not implement BFieldCodec. Moreover, its fields are private,
so an implementation `neptune-core` that agrees with what would be
derived is impossible.

Rather than fixing the issue upstream (by deriving `BFieldCodec` for
`lattice::kem::PublicKey` in twenty-first), the encoder of this commit
simply encodes the public key as bytes and then packs those bytes into
a `Vec` of `BFieldElement`s.

The motivation comes from the need to produce UTXO redemption claims
cryptographically tied to a receiving address. The strategy to achieve
this link consists of producing a particular kind of `Transaction`,
whereby (among other features that set it apart) the tiable receiving
address is a public announcement.

If it should be necessary in the future to `BFieldCodec`-encode
`lattice::kem::PublicKey` objects, then it is preferable to implement
`BFieldCodec` in twenty-first. Such an implementation will lead to
faster code and smaller encodings. However, the branch of the code
base that enables the production and verification of UTXO redemption
claims is not expected to be merged into master, and so modifying
twenty-first (with all the percolating version updates) is not worth
the hassle.