test: Fix off-by-one in RR test-case generator
Fix an off-by-one error in the test-case generator for consistent, synced
removal records, and add a test that triggers the off-by-one error that
was present prior to this commit. It's really important to check these
test-case generators with small AOCL leaf counts as an arbitrary u64 is
with high likelyhood a *really* big number where certain edge case
errors will not materialize.