refactor!(BlockHeader): Replace guesser digest with guesser receiver data
Replace one field that the guesser could set with two fields, one for
the receiver postimage, and one for the lock script hash.

This change enables cold-mining where the secret key capable of spending
guesser rewards lives in a cold vault.

As a consequence of this change, there is no longer a use case for
RawHashLockKey, which means they don't have to be tracked in the wallet,
and that this key type no longer exists.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>