chore: Update TVM benchmark data
The updated benchmark show that the cycle count for hashing the removal
records has decreased by ~25 %. This is due to the better representation
of absolute indices introduced in
e57c6a1debd3c71a694f0481471f0367888b9daa. Note that only the index sets
are hashed, the MMR authentication paths are not.