feat: Automate adding UTXO redemption claims to premine
This commit refactors the premine distribution such that it consist of three
parts:
 1. The original premine distribution, as it is on the legacy chain.
 2. The claims fund for UTXO redemptions.
 3. Exercised claims.

To add new allocations for UTXO redemption claims that come in before relaunch,
simply add a line to `redemption_claims()`.

The amount in the claims fund is set to the sum of all skipped block subsidies
minus the sum of all claims.

Function `Block::premine_distribution` now returns all three classes of
allocations.