refactor!: Add struct and algorithm for memory-hard PoW
Module `pow.rs` contains scaffolding for three interfaces:
 - `preprocess` fills up a buffer of memory
 - `guess` performs one nonce-sampling and produces a `Pow` object (if
   the guess was good)
 - `validate` tests the `Pow` object for validity and for meeting the
   proof-of-work threshold.

The field `nonce` in the `BlockHeader` is now a `Pow`. The nonce lives
on the `Pow`, but the `Pow` also contains other data which is difficult
to sample in a way that satisfies the validity test without storing
large amounts of data in RAM.

Also: implement mast hash for Block, so that the block proof (and not
just the kernel) is hashed to get the block hash. Consequently, you
cannot start guessing before you have a proof.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>