refactor(`Pow`): Modularize
Put functions into the `impl` body of `Pow`. Pass const generic
parameter `MERKLE_TREE_HEIGHT` to `Pow`. This parameter regulates how
much memory is required for proof-of-work. It is possible to supply a
small number for tests. For the consensus, this parameter is specified
by `BlockHeader` at `POW_MERKLE_TREE_HEIGHT` which is defined in terms
of `POW_MEMORY_PARAMETER`, which in turn can be specified later.

Separate out new struct `PowMastPaths`.

Add happy path test.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>