refactor(`Pow`): Tweak definition of leaf
We need to require 41 Tip5 permutations to compute one leaf, but that
does not mean that we need to compute NUM_LEAFS * 41 Tip5 permutations
to compute all of them.

"Buds" are now defined as leafs were: H(commitment || index). Every
leaf (with index i) is now defined as the root of the merkle tree
whose leafs are i..(i+B) wrapping modulo NUM_LEAFS, where B is the
number of buds per leaf. This construction ensures that you only need
log(B) passes over the buffer to generate the leafs.

Drawback: B must be a power of two.