feat: Make PoW preprocessing cancelable
Take an additional argument of type `&dyn Cancelable, which comes with a
`.is_canceled()` member function, which is queried at evenly spaced intervals
throughout the function. If the channel is canceled, abort early and return
the default (empty) `GuesserBuffer`.

This is not a perfect solution yet. On some machines there could be 30s
between cancel points. In the future we want more fine-grained cancel points,
without sacrificing too much performance.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>