ux(`Pow`): Check cancel channel periodically during `preprocess`
Divide the computation in every layer into chunks, and check if the
channel has been canceled between every pair of chunks. As a result,
the `preprocess` step can be aborted in short notice by canceling the
channel. Previously, this step would continue until the layer was
finished, which could take some twenty seconds on a a typical desktop
machine (using present author's machine as a indication).

NB: there is a performance penalty to this periodical check:

| merkle tree height |   original    | periodic interruptions |
|:------------------:|:-------------:|:----------------------:|
|                 25 |   9 s / 2 GB  |           10 s / 2 GB  |
|                 26 |  19 s / 5 GB  |           20 s / 5 GB  |
|                 27 |  39 s / 10 GB |           41 s / 10 GB |
|                 28 |  82 s / 21 GB |           84 s / 21 GB |
|                 29 | 163 s / 42 GB |          171 s / 42 GB |