perf: Drop reliance on `twenty_first`'s Merkle tree
Implement Merkle trees all over again with struct `MTree` living in
module `Pow`. Avoid using `twenty-first`'s Merkle tree for all of
proof-of-work.

The principle motivation is the memory consumption. The existing
Merkle tree builders take a slice of leafs of length N and allocate 2N
`Digest`s worth of memory. In the present context, a) memory is
extremely expensive, and b) the right amount of memory was already
allocated in the bud-to-leaf step. Reusing the already-allocated
memory saves 50% of the peak memory use in `preprocess`.

A secondary motivation is to build in aborts, so that if the guesser
task is be canceled in the course of Merkle tree construction, then
Merkle tree construction halts more or less immediately. This feat is
left to a future commit.

The performance benefit in terms of time was unexpected.

| merkle tree height |   original    |     unsafe    |     mtree     |
|:------------------:|:-------------:|:-------------:|:-------------:|
|                 25 |   9 s / 2 GB  |   9 s / 2 GB  |   8 s / 2 GB  |
|                 26 |  19 s / 5 GB  |  20 s / 5 GB  |  18 s / 5 GB  |
|                 27 |  39 s / 10 GB |  40 s / 10 GB |  37 s / 10 GB |
|                 28 |  82 s / 21 GB |  82 s / 21 GB |  76 s / 21 GB |
|                 29 | 163 s / 42 GB | 166 s / 42 GB | 155 s / 42 GB |