feat(`Pow`): Make Merkle tree construction cancelable
Percolate the cancel channel to `MTree::build_inplace`. Check it
with the same periodicity as in caller function `preprocess`.

Add a test that the process can be aborted. For unknown tokio policy
reasons, that test is a little flaky if the timeout is 1s.

NB: there is a (small) performance penalty when this periodic checking
is on, summarized below.

| merkle tree height |   original    |   mtree (on)  |  mtree (off)  |
|:------------------:|:-------------:|:-------------:|:-------------:|
|                 25 |   9 s / 2 GB  |   8 s / 2 GB  |   9 s / 2 GB  |
|                 26 |  19 s / 5 GB  |  18 s / 5 GB  |  26 s / 5 GB  |
|                 27 |  39 s / 10 GB |  37 s / 10 GB |  38 s / 10 GB |
|                 28 |  82 s / 21 GB |  76 s / 21 GB |  78 s / 21 GB |
|                 29 | 163 s / 42 GB | 155 s / 42 GB | 159 s / 42 GB |