test(mutator_set): Fix flaky test
For some reason the `filter` decoration was not respected, and the
arbitrary derivation failed on the assert changed in this commit -- the
assert verifying that the mutator set has at least as many AOCL leafs as
requested removals. So I rewrote the proptest and associated call to the
arbitrary generator to not need the filter.