test: Fix test assuming unpacked removal records
This test used to work even though a wrong assumption was made, that the
removal records were not packed. The reason was that no authentication
path was present in the removal records list, so packing had no effect.
Now that more UTXOs have been added to the premine (through claims), the
window has slid so much that some removal records have non-empty MMR
auth paths, in which case packing makes a difference.