refactor(main_loop): Avoid relaying empty transactions
Importantly, the transactions are relayed *after* they are processed by the
main loop. In this case, the transaction is processed, but its relay is only
triggered if it is not empty.

The alternative strategy would be to ignore empty transactions in the peer
loop where they come in. Under that strategy there would be no state update
at all, in particular in regards to to the mempool, when an empty transaction
is received.

Under the chosen strategy, nodes who want the mempool state updates can still
trigger them by requesting empty transactions for peers. The rest of the network
does not need to care.