test: re-enable test that first blk-file contains valid blocks
Since the block structure is different in production from what it is
under the test flag, this tests, that checks the validity of the first
~120 mainnet blocks, cannot live under the test flag. So it's added as
an integration test.

Unfortunately some test helper function were needed for this test to
work, and since they can't be under the test flag, they are added as
local functions in the new test. The alternative seems to be to pull the
test helper functions in question out from under the test flag but that
would require new dependencies in production code (namely HTTP client
'clienter') which I think is not preferable. So local functions was the
best compromise I could come up with. Happy to have other people give me
better ideas.