test: Restore state from two blk files
Add an extra blk file moving the state restoration from 127 blocks to
250. This requires CI to download over ~256MB per test run which,
admittedly, is a bit much. The existence of this test does help me sleep
better though, as ensures that the block validation rules (especially
the difficulty mechanism) haven't changed, and that the wallet state is
synced after applying this many blocks.